Change im.keystroke to take additional "keyCode" param.
Most IMs need a universal way to understand a "key" (by code), which is not the
special IM key names (can be retrieved by jscin.get_key_val).

To simplify that and the way we deal with external IMs, add that into
im.keystroke.

Also changed external IM protocol to accept arbitrary commands.