Add "PHRASE_CHARDEF" command to allow having phrases in CIN table.
When this option is turned on, multiple characters in one %chardef mapping won't
be joined as multiple candidates. It'll be kept as-is, which can be used for
phrase based input.

Note gen_inp1 may not work properly. Some features like globing or phonetic
grouping may also fail to work. This option should be turned on explicitly only
on tables designed for phrase.