Change JS scripts into ES6 modules
The JSCin was implemented before ES6 module (import/export) was widely
available so the way to extend modules were poorly implemented.

It is time to change to the better way maintaining JavaScripts in
modules.

Note: Google Drive support is now changed to "always imported, but
the variable in background.js can disable generating the related buttons
in options.js.