Rewrite impl_chromeext (content+background) in ES6 classes
Follows the same approach as croscin.js.

One exception is `ipc.listen`: using => will make the
arguments being an empty object, and not sure why.

The workaround is to keep using function() with `let self=this;`.

Verified that extension still works on non-Chromebooks
(commit text still failed but that's another story).