Change commitText to directly change input value
Previously the Chrome browser supports changing input and textarea
elements' data (value) by dispatching a TextEvent, however the TextEvent
is deprecated by InputEvent, and the InputEvent was just a notification
and will not change the input box contents.

As a result, we now have to directly change the input.value and send out
the InputEvent as notification.