Revise the comments related to accepted_keys
The accepted_keys is very important in IPC based implementations because
we can't wait IPC to finish before we can decide if the keydown event
should be cancelled or not.

The previous implementation builds the accepted_keys using 'key' instead
of 'code', and that became a problem because the KeyboardEvent may
provide a modified key when Shift is hold, and we can't easily create
the mapping. And even worse, some IMs/addons prefer 'key' while other
IMs/addons may prefer 'code' in their logic.

This issue is not fixed yet, but this change revised some incorrect
comments and added more information explaining more details.