Drop selkey2 (unused)
The selkey2 was introduced by commit 62b0f3, reflecting the 'selkey2'
option in gtab (gcin). However, it was implemented only because its
selkey has a fixed length (12 chars) so if the input table needs more
selkey, that will be put to the selkey2 buffer.

In jscin, the selkey is a variable length string so we don't need to
worry about that.