Revise trace and debug messages
The gen_inp1 and gen_inp2 uses lots of 'trace' that was not really
needed.  We should change most of them to debug(), and remove those not
important.

Also, we should never do string + in debug because that will be an extra
overhead even if we don't wont' print the message.