Add new 'config.js' module for cross-extension-pages config
Since Chrome Extension Manifest V3, the background page will be removed,
and the content scripts may be stopped. The new preferred way to sync
across extension content script, option page, and background script
(background service worker) will be `chrome.storage`.

Previously because localStorage can't be shared, the content script
must always do an IPC to figure out if it should start the emulation,
and that takes a long time to load modules and finish.

The new introduced 'chrome/config.js' is a wrapper to the chrome.storage
like persistant data backend. With that we can now easily check if the
emulation should be started in the content script, reducing the initial
time from ~450ms to ~90ms.