cin_parser: Support %ename in multi locales
There are some CIN tables putting a multi-locale string in %ename, for
example:
https://github.com/chinese-opendesktop/cin-tables/blob/master/3corner.cin

Taking that into 'name' will be an error, with the special name added
in to the system and not able to delete it.

The first step is to support parsing that special format.
A future work is to convert %ename to id-friendly strings.