Backup tables from localStorage of Manifest V2 to the storage in V3
Prepare for Manifest V3, we want to backup the user downloaded tables
from localStorage to chrome.storage.local.

Raw is not backed up because from the past of the years, it is very
clear that we rarely see tables that need to re-parsed by the cin/gtab
parsers, and we should drop that in the future.