options: Fix 'newly added table doesn't show in the IME menu'
The extra 'this' caused the options not able to add a new input method
table and have the croscin updated UI correctly.

However if you change the order of the IM immediately, that will fix the
problem even without this patch.

Also fixed that we created the wrong InputMethod list structure. Again,
that will be fixed when you restart the extension.