Fix IM list in options fail to work when the IM ename is not safe for ID
If the user tries to import an IM with special ename (for example those
contains ':'), the options UI will fail and not able to delete the IM
because we used the ename as the ID in jQuery expression.

The right approach is to create an unique ID for each IM. In this patch
we try to turn the ename into a binascii string because that was
supposed to work.