Backup tables in asynchronous, and save as native objects
The chrome.storage provided a "getBytesInUse()" API so we can measure
the real impact to system performance. With that, we found when storing
the largest table (newcj3.cin, 1.6M as UTF-8 text file) in
chrome.storage:

- The native object: 408550 bytes
- JSON stringify'ed: 522422 bytes
- LZString(JSONstr): 373649 bytes

The LZString does have best compression ratio in size.

Time to save:
- LZString+JSON is around 360ms
- Native object takes a long time to write, around 3600+ms.

Time to load:
- LZString+JSON is around 246ms
- Native object is around 139ms

As a result, native objects are fastest in loading, with an acceptable
compression rate. As a result, for MV3 we should move to native objects
in chrome.storage.

And since saving tables now take a very long time, we should let backup
happen in background and not blocking the initialization.