Revise Boshiamy space style
For Boshiamy (space_style=1), the implementation sets both SELKEY_SHIFT
and SPACE_AUTOUP to work with the official gtab (selkeys=1~9, but the
first candidate should be space). However, some Boshiamy tables have 0-9
as selkeys so actually 0 = SPACE = first candidate - and this is how
some other platforms work. Having both SELKEY_SHIFT + SPACE_AUTOUP will
change the key ordering in this case.

And just for reference - in Dayi's case, they use '[]-= as 1-5
selection, with space as the first candidate; and the documents clearly
said you have 6 candidates that can be directly selected without holding
Shift.

The solution is to decide what key to shift based on the selkey.
Note this changes currently only applies to gen_inp2.