Find available modules when the default does not exist
Just like config.InputMethods(), when we can't find a module then we
should always choose the only registered module.

Note: This is based on the assumption that all registered modules are
(almost) compatible with each other. We need to revisit this again if we
support other modules (like CrExtInp) again.