Add a new placeholder input method 'en-US'
The en-US is an empty CIN table with nothing defined. The IME will
simply bypass all key strokes to the system.

This CIN file is created for a better user experience, so when the users
first time install the JsCIN, they can choose and activate only the IMEs
they need/want, and no need to waste the time and space for the built-in
IMs that they had no plan to use forever. (Each IM table takes 100~300k
so that's large).

The cin_parser must be patched otherwise a single '%selkey' will be
treated as 'true' and causing the input module to fail.