Remove the raw tables
The raw tables in localStorage was designed long ago that we have the
concern that the parser was not buggy, so saving only the parsed data is
not good; with raw data we can parse again and get the better results
when updated.

However, practically this is not going to happen because
- If a table has some serious problem, users tend to remove it
- It is not clear which type we should choose if we want to parse again
- The stored data was pretty close to the real CIN files
- The gtab parser is also stable
- For remote downloaded tables, they can always reload using URL

As a result, it seems easier to just drop the 'raw' feature that will
also save more space in the storage.