jscin/ime.js: A new manager for all input methods in jscin.
The jscin.js worked like a set of tools to maintain a set of input
method related components, but not putting them all in an organized way.

The newly implemented ime.js (from scratch), not only abstracted and
simplified the access to the underlying storage system (by
jscin/storage.js), supports async I/O, but also made the interface
pretty simple and memory efficient (load only when needed).
It is also the key of the migration to the new Manifest V3.