Stop backupTables
We will do migration directly using the localStorage, so there is no
need to have the backup tables. Also in migration, we are not deleting
the old table so in ime.js we need to delete both the new and old
tables; otherwise deleted tables will show again after reboot (another
migration).