Introduce 'DefaultModule' for config and revise the implementation
We wnt to change the DefaultModule to be a Config property. However,
that also implies moving from jscin (input methods environment/manager)
to croscin (input methods engine). Currently this is not easy because
jscin controls the registration of input methods and croscin can't pass
the config values to it. So we only updated the new `ime.js` for the new
design (taking module as an param to activateInputMethod).

The 'default' (first registered) module is renamed to 'fallback' to
prevent confusion.