Unify cinParse return variables
cinParse returns two variables, and previously they were mixed as
'result[2]', 'data', ... etc.

Now everyone calls it as `[success, result]` so we know the result is
the error message when success if false, or a structure with [data,
metadata]. The caller should find them as 'result.data' instead of
'data.data'.