Change the 'table.setting' to 'table.type'
It is more clear to name it 'type', with the 'types.json' indicating the
contents of the type data. Also, the 'options' inside 'type' is renamed
to 'cin' since it is exactly the CIN commands.

Note 'type' is currently not really implemented and supported by jcsin/
core except the 'type.cin'. We should consider renaming 'type' to
'extra' and handle the 'cin' options in a different way (leaving that to
croscin and other engines to handle).