Add an FNV-1A hash to the table name
Currently when the ename (actually still a label) is the same, the jscin
will only allow one table to exist. However, it is possible that we may
want to differentiate the tables with the same ename but from different
sources, for example the builtin table and the downloaded table.

Let's add a hash of the URL to solve this problem.