background: Show ChromeOS welcome only when croscin is ready
The croscin now takes a long time to initialize, and if we start the
option even before that then the tables (e.g., the built-in tables)
will be even not ready.

Since the constructor can't be async, we now let the caller invoke
Initialize so they can do await to start when ready.