Add the new emulation: iframe based implementation
The IPC implemenations are slow and complicated, also not friendly for
the new manifest V3 service_worker design. Here we have introduced a new
implementation 'iframe' that shows the UI as iframe, but only sending
the UI event to it and not introdcing new commands / events to the
`chrome.input.ime` interface.

Note the menu was not fixed yet.