Move iframe related files to iframe/ folder.
The iframe implementation actually need a few files: the IME running in
the content script context (e.g., `iframe.js`), the ime_panel in iframe.

To prevent confusion that `iframe.js` itself is not iframe, a new folder
'iframe' is created and `iframe.js` is renamed to `content.js`.

Changed all references to the iframe related files.