Disable the 'all_frames' setting in manifest
In the current implementation of crext emulation, the content script
will send the message to ALL extension pages. The extension components
like ime_panel will check if that's from the same tab, but if a page has
two iframes (each with INPUT or TEXTAREA), then we will have two
`content.js` instances, and two instances of `ime_panel.js`s, causing
endless events. This may be solved by adding a seed to the `content.js`,
use runtime.connect to find out a port, or change the iframe to be
single instance per page. However we don't have a good solution now, so
let's disable the all_frames for now.