menu: Add and use the new event 'MessagePopup'
The `chrome.input.ime` does not have any events that will request to
update the menu. However, in crext emulation mode the menu is provided
by the browser action popup so it need an explicit event notifying the
croscin to resend the menu info. As a result, we have to add a new event
'MessagePopup'. Previously this was done by Activate, but the Activate
actually has more special meaning and should not only update the menu
otherwise it would cause other issues for example the real Activate
won't initialize the composition and candidates window properly.