Show error message on input error
Add an IMKEY_ERROR to notify the croscin that there is some error in the
input data, and let the croscin to display the error message in the
auxiliary text area. This is important when {SPACE,AUTO}_RESET is turned
on.

This is part of the user experience improvement for issue #37.