gtab_parser: issue #39: Support phrases in GTAB
GCIN gtab tries to store mapped characters as UTF8 in 4 bytes. For
phrases or single byte characters, that must be converted to the phrases
list after the items list (e.g. chardef). The gtab_parser didn't support
phrases so we failed to parse some GTAB like vocabulary or array_phases.