gen_inp2: Fix: empty composition still shows candidates
In Array30 with partial_patch turned on, type a few characters and then
keep pressing BackSpace, then we'll see the candidates showing 1-9, each
labeled with 1-9 (actualy the first page of partial matching the whole
table with empty input).

The main reason is in PrepareCandidates, we do try to find candidates
even if the composition is empty (also the partial match didn't reject
empty input). And in ClearCandidates we didn't touch mcch (only in
UpdateCandidates) and that should be fixed as well.