options: Fixed "adding a table will see duplicated items in the IM list"
This issue was started when we introduced hash to the table names.
The save_name only works in 'reload' but no the rest, so here we want to
remove the save_name pasing and always try to right (expected) name.
Reload should still have the right name assuming the %ename in file is
the same (and if not, it seems no harm to just have a new entry).