gen_inp2: Limit glob to only one page in auto_compse preview
In composition (before conversion) stage, there is no key to switch
between multiple pages; so unless we decide to also support cycling the
candidates page in composition mode, limit the preview to only one page.

Note: as an example, Boshiamy and IMs that used %SPACE_AUTOUP will be
actually not able to cycle the candidates after conversion by SPACE
because that is their commit key.