ime: Fixed "CrossQuery failed to work with built-in tables".
If we set the CrossQuery to build-in tables and haven't activated that
input method since the extension started, the addon won't be able to
load the table.

This is because loadTable need an url to load when it's the built-in
tables. As a result, we should look up the TableInfo in loadTable to get
the URL and then load the built-in tables.