Add 'page of candidates' to the auxiliary text
In partial-match or globbing modes, we will want to know how many pages
of candidates we are cycling. The `chrome.input.ime` seems not doing the
page processing well so we have to do it inside the auxiliaryText.

The cross-query addons is also revised to follow the new protocol for
displaying messages in the auxiliaryText field.