gen_inp2: Change the match limit to be the same in different states
In most IMs, they will see the auto compose preview and then press
SPACE. The experience is better if we always do "when multiple pages,
SPACE means moving to the next page".

Array30 is still not following this (because it has the extra %quick)
and can be improved in the future.

Also enlarged the default max page from 10 to 20.