Fix "Turn on/off SPACE_AUTOUP several times and 1st SEL becomes SPACE"
This is found on IMs with SPACE_AUTOUP+SELKEY_SHIFT. The quirks will try
to modify the selkey to start with SPACE. However, since the table was
cached in `jscin.loadTable`, the modified result will remain in the
system. And even worse, because %SELKEY_SHIFT was kept in the table (for
opts to track), everytime we reload this IM, its selkey will keep
changing.

The correct fix is to clone the CIN before applying quirks. Also fixed a
bug that 'jscin.freeTable' never worked in the past (although we now
choose to dupe instead of free).