croscin: Track and use the last activated input method
Previously, selecting a non-default input method will be noly in the
runtime variables and will be lost when the extension service worker
become deactivated. This is actually easily reproduced by:

 1. Select the 2nd IM
 2. Ctrl-Space to deactiavte the IM (also stopped heartbeat)
 3. Wait for a few mins, check from the extension page that service work
    become deactivated
 4. Ctrl-Space again, the activated IM is the default (1st).

And this would be annoying for peoeple who wanted to temporarily switch
to the non-default IM. As a result, we should track that as well.
Although the default IM will be still selected if you make changes in
the options page.