croscin: Improve Raw mode behavior on switching
When the user switches from the normal mode to raw mode (with some
keystrokes still in the composition buffer), we are expecting they have
done some input in the wrong mode - e.g., those were english instead of
raw typing. As a result, we should commit the 'raw composition' in
composition state, or the 'first candidate' in candidate selection
state.