gen_inip2: Process SPACE as selection key in candidates mode
This is a bug introduced by commit 64f063f. For IMs really used SPACE
(not 0) as the selection key (in this test, dayi3p), you can't actually
select the first candidate even if there's only one page. And in fact
you can never select the first candidate because when the SPACE_AUTOUP
is not set. As a result, we need to allow SPACE as selkey, and for such
IMs they have to first press SPACE for conversion, then arrow to cycle
for the other pages.