ime: Change SPACE_AUTOUP to default true
This is to fix the `gen_inp_test.js` failure introduced by commit
64f063f. Aparently most IMs do expect 'SPACE after convert' should
commit the first word, so we do want to make it default.

Also revised the comments for Phonetic quirks to make it clear that
AUTO_COMPOSE is a problem when we enable SPACE_AUTOUP.

For Array30, as verified on OpenVanilla (which does not have the option
to turn on/off SPACE_AUTOUP), the behavior is the same so we are
following the same default.