gen_inp2: Allow Array30 to explicit 'Convert' in %quick mode
The commit 77acc57 changed %SPACE_AUTOUP to default, and that will
actually create a problem in Array30 when we have a keystoke that
 - is covered by %quick
 - has multiple candidates within one page

One example is 'p'.

From %SPACE_AUTOUP, the candidates were already displayed so we should
be good to just commit the first candidate. However, the first SPACE
actually triggered changing from %quick (overrides) to the standard key
table so we should enter 'candidate selection' stage and wait for one
more key input. In other words, we have to track "if previous candidates
were overriden or not". This also aligns with OpteVanilla behavior.