gen_inp2: Ensure commit "first of the current page" instead of "first candidate"
The IM will usually want to commit "the first candidate of the current
page", but CommitText(ctx, 0) will actually commit the "first of the
whole candidate list", so let's create a new CommitFirst for "first of
the current page".

This is required for follow up change to support "multiple candidate
pages in the composition mode".