gen_inp2: Support page navigation in compostion mode
For IMs like Boshiamy that prefers SPACE=[1st candidate], they can't
do the page navigation in STATE_CANDIDATES so we have to also support
moving the start index when multiples are available.

And when 'first press pageDn in composition state', we have to shift the
state and move the index at the same time; as a result, we have to keep
the start_index unchanged in ShiftState call.