gen_inp2: Space for IMs with %quick always need an explicit SPACE
Prevoiusly we tried to track the override status and allow the SPACE to
behave like IMs without %quick when the keystrokes were outside of the
%quick (e.g., 3rd or 4th in Array30). However, checking the '....'
behavior, it seems like on OpenVanilla the first SPACE puts the first
candidate into the composition area while the candidates are still
selectable. That is also better aligned with AUTO_COMPOSE=false.