ime_api: Stop setting body css in webpage implementation.
The `webpage.js` used set body CSS because the shared `ime_panel.css`
set body to be tranparent, and that only makes send in the iframe-based
IME panel.

To fix that, we move the transparency to be only for class imeBody; and
only the `ime_panel.html` inside iframe has that class name set.