ime_api: Hide candidate window properly
We want to also check and verify the candidate window `visible` property
behavior, so that should be applied to the `testarea`. However, it's not
easy to use in the IME Frame so we will keep current behavior (only
change the candidates list) for that.

Also fixed a bug that TestAreaPanel was not a block.