quirks: Fixed: issue #41: SELKEY_SHIFT from saved opts didn't work
If the input method table doesn't choose %space_style=1 and the quirks
didn't detect it as Boshiamy or Dayi, then the SELKEY_SHIFT won't work
even if the user saves that in the Options opts list, because the
options were applied after all quirks were set.

We need a new post-quirks to adjust the settings with the final opts.
Also revised the messages so it's more aligned with the behavior.