quirks: Revise SPACE_AUTOUP comments & description
We want to align more with gcin space_style=1 behavior.

In gcin, when space_style that actually implies SELKEY_SHIFT for most
IMs, but for tables with %quick, it will keep selkey unchanged, but
also enforce only showing one page for candidates (no multi-page).

In JsCIN we do have an explicit SELKEY_SHIFT option so we should not
make lots of exceptions.