messages & test: Use 'cangjie' as the formal name
The modern way of calling the `cj.cin` input method is 'cangjie'.
The legacy name is 'changjei'. And 'cangjei' is rarely seen.

We don't want to change 'changjei' because it's defined in the table we
use, but for showing as input method type, 'cangjie' is much better and
we should never use 'cangjei' in the system.