addon_related: Decide if shift is required without AUTO_COMPOSE
Previously we decide if the Shift is required by checking AUTO_COMPOSE
states, but that is not valid if the user choose to turn off the
AUTO_COMPOSE opt.

The right way to check that is to look at %chardef and %selkey, allowing
only if the %chardef never starts with %sekley (or if the %selkey is
mapping only to itself - which should cause confusion in the composition).